moved devDungeon logic to framework to support blockly
