+++ blockly/build.gradle
@@ -8,7 +8,6 @@ repositories {
 
 dependencies {
   implementation project(':dungeon')
-  implementation project(':devDungeon')
 
   // ANTLR version 4 for DSL Grammar
   antlr supportDependencies.antlr
@@ -26,7 +25,6 @@ sourceSets.test.java.srcDirs = ['test/']
 processResources {
     from new File(project(':game').projectDir, '/assets')
     from new File(project(':dungeon').projectDir, '/assets')
-    from new File(project(':devDungeon').projectDir, '/assets')
     duplicatesStrategy = DuplicatesStrategy.EXCLUDE
 }
 
@@ -45,12 +43,11 @@ tasks.register('runBlockly', JavaExec) {
 }
 
 tasks.register('buildBlocklyJar', Jar) {
-    dependsOn ':game:jar', ':dungeon:jar', ':devDungeon:jar'
+    dependsOn ':game:jar', ':dungeon:jar'
     archiveBaseName = 'Blockly'
     from sourceSets.main.output
     from project(':game').sourceSets.main.output
     from project(':dungeon').sourceSets.main.output
-    from project(':devDungeon').sourceSets.main.output
     from {
         configurations.runtimeClasspath.collect { it.isDirectory() ? it : zipTree(it) }
     }
@@ -65,7 +62,6 @@ tasks.register('buildBlocklyJar', Jar) {
     into('assets') {
         from new File(project(':game').projectDir, '/assets')
         from new File(project(':dungeon').projectDir, '/assets')
-        from new File(project(':devDungeon').projectDir, '/assets')
     }
 
     exclude('META-INF/*.SF', 'META-INF/*.DSA', 'META-INF/*.RSA')
