+++ devDungeon/src/utils/EntityUtils.java
@@ -1,8 +1,6 @@
 package utils;
 
 import contrib.components.HealthComponent;
-import contrib.entities.SignFactory;
-import contrib.utils.components.skill.SkillTools;
 import core.Entity;
 import core.Game;
 import core.components.PositionComponent;
@@ -11,7 +9,6 @@ import core.level.elements.ILevel;
 import core.level.elements.tile.ExitTile;
 import core.level.utils.Coordinate;
 import core.utils.IVoidFunction;
-import core.utils.MissingHeroException;
 import core.utils.Point;
 import core.utils.components.MissingComponentException;
 import entities.MonsterType;
@@ -27,9 +24,9 @@ import task.tasktype.Quiz;
 
 /**
  * EntityUtils is a utility class that provides methods for spawning entities in the game. It
- * contains methods for spawning monsters, signs, torches, levers, and other entities. It also
- * contains methods for teleporting entities to specific positions and retrieving the position of
- * the hero in the game.
+ * contains methods for spawning monsters, torches, and other entities.
+ *
+ * @see contrib.utils.EntityUtils
  */
 public class EntityUtils {
 
@@ -96,44 +93,6 @@ public class EntityUtils {
   }
 
   /**
-   * This method is used to spawn a sign entity in the game at a given position. It uses the
-   * SignFactory class to create a new sign with the provided text and title. The sign is then added
-   * to the game. If an IOException occurs during the creation of the sign, it is caught and a
-   * RuntimeException is thrown.
-   *
-   * @param text The text to be displayed on the sign.
-   * @param title The title of the sign.
-   * @param pos The position where the sign should be spawned.
-   * @param onInteract The action to perform when the sign is interacted with. (sign, whoTriggered)
-   * @return The spawned sign entity.
-   * @throws RuntimeException if an error occurs while spawning the sign.
-   * @see contrib.entities.SignFactory#createSign(String, String, Point, BiConsumer) createSign
-   */
-  public static Entity spawnSign(
-      String text, String title, Point pos, BiConsumer<Entity, Entity> onInteract) {
-    Entity sign = SignFactory.createSign(text, title, pos, onInteract);
-    Game.add(sign);
-    return sign;
-  }
-
-  /**
-   * This method is used to spawn a sign entity in the game at a given position. It uses the
-   * SignFactory class to create a new sign with the provided text and title. The sign is then added
-   * to the game. If an IOException occurs during the creation of the sign, it is caught and a
-   * RuntimeException is thrown.
-   *
-   * @param text The text to be displayed on the sign.
-   * @param title The title of the sign.
-   * @param pos The position where the sign should be spawned.
-   * @return The spawned sign entity.
-   * @throws RuntimeException if an error occurs while spawning the sign.
-   * @see #spawnSign(String, String, Point, BiConsumer) spawnSign
-   */
-  public static Entity spawnSign(String text, String title, Point pos) {
-    return spawnSign(text, title, pos, (e, e2) -> {});
-  }
-
-  /**
    * Spawns a bridge guard at the given position and adds it to the game. The bridge guard neutral
    * NPC that gives the player a series of {@link task.tasktype.Quiz quizzes} to solve. The bridge
    * guard is created using the {@link entities.MonsterType#BRIDGE_GUARD BRIDGE_GUARD} monster type.
@@ -149,40 +108,6 @@ public class EntityUtils {
     Entity bridgeGuard = MonsterType.createBridgeGuard(pos, quizzes, onFinished);
     Game.add(bridgeGuard);
     return bridgeGuard;
-  }
-
-  /**
-   * This method is used to spawn a Lever entity in the game at a given position. It uses the
-   * LeverFactory class to create a new Lever with the provided onInteract action. The Lever is then
-   * added to the game. If an IOException occurs during the creation of the Lever, it is caught and
-   * a RuntimeException is thrown.
-   *
-   * @param pos The position where the Lever should be spawned.
-   * @param onInteract The action to perform when the Lever is interacted with. (isOn, lever,
-   *     whoTriggered)
-   * @return The spawned Lever entity.
-   * @throws RuntimeException if an error occurs while spawning the Lever.
-   * @see entities.LeverFactory#createLever(Point, ICommand) createLever
-   */
-  public static Entity spawnLever(Point pos, ICommand onInteract) {
-    Entity lever = entities.LeverFactory.createLever(pos, onInteract);
-    Game.add(lever);
-    return lever;
-  }
-
-  /**
-   * This method is used to get the coordinates of the hero in the game. It uses the SkillTools
-   * class to get the hero's position as a point and then converts it to a coordinate. If the hero
-   * is missing, it catches the MissingHeroException and returns null.
-   *
-   * @return Coordinate of the hero's position. If the hero is missing, returns null.
-   */
-  public static Coordinate getHeroCoords() {
-    try {
-      return SkillTools.heroPositionAsPoint().toCoordinate();
-    } catch (MissingHeroException e) {
-      return null;
-    }
   }
 
   /**
@@ -339,72 +264,5 @@ public class EntityUtils {
    */
   public static Entity spawnBoss(MonsterType bossType, Coordinate levelBossSpawn) {
     return spawnBoss(bossType, levelBossSpawn, (e) -> {});
-  }
-
-  /**
-   * Teleports the hero to a specified point in the game.
-   *
-   * <p>This method retrieves the hero entity from the game and calls the teleportEntityTo method to
-   * change the hero's position. If the hero entity is not present (which can happen if the hero has
-   * fallen into a pit), the method does nothing.
-   *
-   * @param point The point to which the hero should be teleported.
-   */
-  public static void teleportHeroTo(Point point) {
-    Entity hero = Game.hero().orElse(null);
-    if (hero == null) {
-      return;
-    }
-    teleportEntityTo(hero, point);
-  }
-
-  /**
-   * Teleports an entity to a specified point in the game.
-   *
-   * <p>This method changes the position of the given entity to the specified point. It does this by
-   * fetching the PositionComponent of the entity and setting its position to the given point. If
-   * the entity does not have a PositionComponent, a MissingComponentException is thrown.
-   *
-   * @param entity The entity to be teleported.
-   * @param point The point to which the entity should be teleported.
-   * @throws MissingComponentException if the entity does not have a PositionComponent.
-   */
-  public static void teleportEntityTo(Entity entity, Point point) {
-    PositionComponent pc =
-        entity
-            .fetch(PositionComponent.class)
-            .orElseThrow(() -> MissingComponentException.build(entity, PositionComponent.class));
-    pc.position(point);
-  }
-
-  /**
-   * Retrieves the current position of the hero in the game.
-   *
-   * <p>This method retrieves the hero entity from the game. If the hero entity is not present
-   * (which can happen if the hero has fallen into a pit), the method returns null.
-   *
-   * @return The current position of the hero, or a null value if the hero is not present.
-   */
-  public static Point getHeroPosition() {
-    return Game.hero()
-        .map(
-            e ->
-                e.fetch(PositionComponent.class)
-                    .orElseThrow(() -> MissingComponentException.build(e, PositionComponent.class))
-                    .position())
-        .orElse(null);
-  }
-
-  /**
-   * Retrieves the current coordinates of the hero in the game.
-   *
-   * <p>This method retrieves the hero entity from the game. If the hero entity is not present
-   * (which can happen if the hero has fallen into a pit), the method returns null.
-   *
-   * @return The current coordinates of the hero, or null if the hero is not present.
-   */
-  public static Coordinate getHeroCoordinate() {
-    Point heroPos = getHeroPosition();
-    return heroPos == null ? null : heroPos.toCoordinate();
   }
 }
