+++ dungeon/src/contrib/utils/ICommand.java
@@ -1,10 +1,10 @@
-package utils;
+package contrib.utils;
 
 /**
  * ICommand is an interface that represents a command that can be executed and undone.
  *
- * @see entities.LeverFactory LeverFactory
- * @see systems.LeverSystem LeverSystem
+ * @see contrib.entities.LeverFactory LeverFactory
+ * @see contrib.systems.LeverSystem LeverSystem
  */
 public interface ICommand {
